transition all previous content to new jsx files
