+++ src/css/main.scss
@@ -123,7 +123,7 @@ article {
 .main-menu li {
   float: left;
   padding: 20px;
-  border: 1px solid black;
+//  border: 1px solid black;
 }
 
 .main-menu-item a {
