+++ src/js/components/Hero.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+export class Hero extends React.Component {
+
+	render() {
+		return (
+			<div className="tagline">
+				<h1>Can you really do Church IT effectively?</h1>
+				<h1>Anything is possible.</h1>
+				<h1>Come on, we'll show you.</h1>
+			</div>
+		);
+	}
+}
